query-genie-backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
docopt==0.6.2
et_xmlfile==2.0.0
ete3==3.1.3
fastapi==0.115.8
future==1.0.0
h11==0.14.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
pymongo-schema==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
XlsxWriter==3.2.2
xlwt==1.3.0
